refactor(Test): extract verify tests to own test file
